refactor(providers): descriptors build session configs; slice dispatch via settingsSliceKey
Move the 8 createXxxSessionConfig builders from settingsStore into their
provider descriptor classes as buildSessionConfig; OpenAI/Compatible share a
module-level buildOpenAISessionConfig helper (Compatible inherits). Turn
createSessionConfig, getCurrentProviderSettings, useCurrentTurnDetectionMode,
and LanguageSection's duplicate slice map into registry-driven shells keyed by
settingsSliceKey. Unknown providers now fail loud via getDescriptor instead of
silently falling back to the OpenAI slice.

settingsStore.test.ts gains an environment mock (mirroring
descriptorRegistry.test.ts) so the env-gated VOLCENGINE_AST2 descriptor is
registered in jsdom; assertions are unchanged.
